+++ authorityspoke/utils/marshmallow_oneofschema/__init__.py
@@ -1,1 +1,0 @@
-"""Schema that multiplexes other schemas based on object type."""